personal_wellness_tracker_backend/pyproject.toml
[tool.poetry]
name = "personal-wellness-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for Personal Wellness Tracker application"
authors = ["Alifriduwan <6510110565@email.psu.ac.th>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
pydantic = "^2.11.9"
fastapi-cli = "^0.0.11"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.10.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
